Udacity Data Analyst Nanodegree Data Visualization Project
This project uses the Titanic data set to create data visualizations
using dimple.js primarily. It includes an initial and a final html
file showing the visualized charts, the raw data csv files, a link to
an older Titanic data analysis project from which the charts data was
sourced and a README file which contains the project report.

Signed-off-by: Smita Vaidya <smita1005@gmail.com>